app/requirements.txt
fastapi
uvicorn
gunicorn
sentence-transformers
torch
pillow
opencv-python
scenedetect
python-multipart
faiss-cpu
